refactor: add util to determine element visibility
